fix pipeline caching bug caused by bool
the rendergraph pipeline desc had a field `bool is_transparent` that was
set in a struct literal. however i guess `bool` is not a standardized
size, and the remained of struct padding had trash memory that was not
properly zerod. That struct field is now a u32, and we use a helper to
set it.

the trash memory in the padding was incorrectly causing a cache miss,
and randomly triggering the creation of new renderpipelines.

nasty bug.