8-byte-align CQ commands; fix Ben transparency
All command-queue commands are now aligned to 8bytes to avoid
undefined behaviour warning from xcode, also byte-aligns memory
accesses.

Fix horrible transparency bug caused by a bitwise OR between the
u64 sortkey and the R_ID material_id; when the material_id was
sufficiently high, for some reason it was cast into a very large
value and trashed the sortkey. probably bc the bitshift
turned it into a signed negative value. now we OR the sortkey with
a u64 material_mask.